requirements.txt
pymupdf>=1.23
pydantic>=2.0
PyYAML>=6.0
loguru>=0.7
pytest>=7.4
